requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For parquet file support

# Scientific computing
scipy>=1.10.0
filterpy>=1.4.5  # Kalman filter for AFNS state-space model

# Machine learning
scikit-learn>=1.3.0
hmmlearn>=0.3.0  # Hidden Markov Models for regime classification

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

